requirements.txt
fastapi[standard]
uvicorn
pydantic
pymysql
SQLalchemy
mysql
python-dotenv
python-multipart
aiosmtplib
python-jose
pydantic[email]
passlib[bcrypt]
python-multipart
jwt
pyjwt
alembic